Remove unused component Switcher, move HeaderMenuButton to Navbar folder
