--- src/components/carbon/Switcher/index.js
@@ -1,3 +1,0 @@
-import Switcher from './Switcher';
-
-export default Switcher;
